Changed DWORD to unsigned long to allow for pointers

+++ controller/source/engine/timer.cpp
@@ -124,9 +124,10 @@ void DelayLine::ReceiveClockTiming(void* pParam)
 }
 
 
-bool DelayLine::GoOneShot(DWORD Param)
+bool DelayLine::GoOneShot(unsigned long Param)
 {
 
+  
   bool ret = false;
   //get the time
   DWORD CallbackTime = Scheduler::getTime() + wInterval;
